[ADD] Read_images function
The `read_images` function takes the folder path where images are, the image format and the path of the file with the labels of the images. We suppose images are separated in different folders named with the labels. 

This functions returns a string array with the label names and a dictionary with the images for each label.