[MODIFY] Main file accepts new arguments.
These arguments indicate if a device is running as a edge, fog or cloud device. Besides, one of them will contain the next device direction to send the inference results. The last argument contains a threshold value which indicates the minimum value to be accepted as a correct inference result in the classification output of the device, but it is only needed if the `next-device` argument is specified.